Make function location only the start of the function.
Not sure if it should be the first statement instead... but this will do for now. This basically prevents non-code lines from being marked as executed in the function. If the function body is marked by this then the execution count for every line in the function was _at least_ the number of invocations of the function – which isn't necessarily true.