Strip hardcoded analyze tags.
Previously there were `statements`, `branches`, etc. in the analysis results; now all that remains is whatever tags are present in the instrumented file. This means if there are no branches then there will be no `branch` entry in the analysis. The names of the tags are singular, so the resulting object has singular keys.