Add better support for `return` statements.
Previously the statement:

```javascript
return foo;
```

was instrumented as:

```javascript
++marker;
return foo;
```

and acquired the location of the return statement. Now it looks like:

```javascript
return ++marker, foo;
```

and it acquires the location of foo. This (at least) fixes cases where returning JSX results in the JSX being marked as un-run even though is clearly is.