Detect partial object/array execution.
This is a bit of an esoteric edge case, but an interesting one nonetheless. Normally `adana` will only instrument the creation of the object itself, not the actual statements that make up the object. So:

```javascript
{
  foo: 5,
  bar: 7,
}
```

Only gets one marker for the whole object (that also happens to span those 4 lines). Now each part of the object is instrumented and if, for some reason, a later property is not actually computed (say due to an exception that gets caught), this is now accurately reported. The same is true for arrays.