Check for destructuring patterns.
Marker injection should not be done for destructuring patterns, since they neither make any sense nor do they compile properly.

Previously there was no logic to detect the fact that in:

```javascript
const x = ({ x: y }) => y;
```

The identifier `y` was not a reference but a declaration. So now destructuring patterns are simply ignored. Test added to verify this behavior for objects and arrays.