Fix invalid conditional group state.
The majority of branches detect the groups they belong to; that is to say if you have:

```javascript
if (foo) {

} else if (bar) {

} else {

}
```

The several branches in there all belong to the same logical group. The same effect occurs for logic statements, tenary ifs and so on. The group detection logic, however, did not take into account adana generated nodes and so compound conditionals with adana generated nodes would result in key errors.

This is now fixed and checks have been added to ensure the correct groups are calculated for the majority of cases. Logic statements still need work, simply because of how they are rewritten for instrumentation.