Replace file hash with file source.
The `hash` value was originally there to ensure you weren't comparing coverage between files you should not have been; having the raw source also achieves this, with the upside of being able to link the coverage data to exact lines in the source. The downside is that it bloats the coverage file and could have potentially sensative information in it.